feat: definition of the final code versions
